requirements.txt
google-generativeai
GitPython